\name{arp23}
\alias{arp23}
\docType{data}
\title{Constituent members of Protein Complex Arp 2/3}
\description{
  A character vector of the constituent members of Arp 2/3 given by the
  common names.
}
\usage{data(arp23)}
\format{
  The format is:
 chr "arp23"
}
\examples{
data(arp23)
}
\keyword{datasets}
