\name{arp23Orf}
\alias{arp23Orf}
\docType{data}
\title{Constituent members of Protein Complex Arp 2/3}
\description{
  A character vector of the constituent members of Arp 2/3 given by the
  systematic gene names
}
\usage{data(arp23Orf)}
\format{
  The format is:
 chr "arp23Orf"
}
\details{
}
\examples{
data(arp23Orf)
}
\keyword{datasets}
