\name{arp23Y2HG}
\alias{arp23Y2HG}
\docType{data}
\title{A graph of ARP 2/3 containing only Y2H verified interactions}
\description{
  An instance of the class graph for the protein complex Arp 2/3's Y2H
  verified edges.
}
\usage{data(arp23Y2HG)}
\format{
  The format is:
 chr "arp23Y2HG"
}
\examples{
data(arp23Y2HG)
}
\keyword{datasets}
