\name{cfia}
\alias{cfia}
\docType{data}
\title{Constituent members of Protein Complex Cleavage Factor IA (CFIA)}
\description{
  A character vector of the constituent members of Cleavage Factor IA
  given by the common names.
}
\usage{data(cfia)}
\format{
  The format is:
 chr "cfia"
}
\examples{
data(cfia)
}
\keyword{datasets}
