\name{cfiaOrf}
\alias{cfiaOrf}
\docType{data}
\title{Constituent members of Protein Complex Cleavage Factor IA (CFIA)}
\description{
  A character vector of the constituent members of Cleavage Factor IA given by the
  systematic gene names
}
\usage{data(cfiaOrf)}
\format{
  The format is:
 chr "cfiaOrf"
}

\examples{
data(cfiaOrf)
}
\keyword{datasets}
