\name{dataS}
\alias{dataS}
\docType{data}
\title{A character matrix containing the source data for the ScISI}
\description{
  This data object is a matrix that gives the description of each of the
  data repositories used to generate the ScISI. 
}
\usage{data(dataS)}
\format{
  The format is:
 chr "dataS"
}

\examples{
data(dataS)

}
\keyword{datasets}
