\name{Desc}
\alias{Desc}
\alias{Desc,yeastData,character-method}


\title{A method to return a description of a protein complex}

\usage{
Desc(object, name)
}
\arguments{
  \item{object}{An instance of a subclass of \code{\link{yeastData}}.}
  \item{name}{A \code{character}. It is the name of the protein complex
    with respect to the bi-partite graph incidence matrix}
}

\description{
  The \code{Desc} method takes in the protein complex name of a bi-partite
  graph incidence matrix (usually an ad hoc name) and returns a
  description of that protein complex.

  The object to be referenced is an instance of the class
  yeastData.
  
}
\keyword{classes}

