\name{eAt2}
\alias{eAt2}
\docType{data}
\title{A file containing edge attributes}
\description{
  This file contains edge attributes used in the rendering of the graph
  ARP 2/3 in the ScISI.Rnw paper.
}
\usage{data(eAt2)}
\format{
  The format is:
 chr "eAt2"
}
\examples{
data(eAt2)

}
\keyword{datasets}
