\name{egEBI16112}
\alias{egEBI16112}
\docType{data}
\title{A graph example mapping an IntAct ID to Systematic Gene Names}
\description{
  This file is an instance of the class graph showing the mapping of the
  IntAct ID EBI-16112 to multiple Systematic Names.
}
\usage{data(egEBI16112)}
\format{
  The format is:
 chr "egEBI16112"
}

\examples{
data(egEBI16112)

}
\keyword{datasets}
