\name{expStats}
\alias{expStats}
\docType{data}
\title{A data file containing the experimental statistics}
\description{
  This file is a matrix that details the number of protein complexes
  found within each data repository (MIPS, GO, Gavin, Ho, Krogan) as
  well as the number of genes present (pre-processing).
}
\usage{data(expStats)}
\format{
  The format is:
 chr "expStats"
}
\examples{
data(expStats)

}
\keyword{datasets}
