\name{gavin2mergeMG}
\alias{gavin2mergeMG}
\docType{data}
\title{A data file containing the pre-merged Gavin to (merged) MIPS-GO data}
\description{
  This file contains the pre-processed Gavin data before merging to the
  joint MIPS-GO interactome.
}
\usage{data(gavin2mergeMG)}
\format{
  The format is:
 chr "gavin2mergeMG"
}
\examples{
data(gavin2mergeMG)

}
\keyword{datasets}
