\name{ho2mergeMGG}
\alias{ho2mergeMGG}
\docType{data}
\title{A data file containing the pre-merged Ho to the (merged)
  MIPS-GO-Gavin data}
\description{
  This file contains the pre-processed Ho data before merging to the
  joint MIPS-GO-Gavin interactome.
}
\usage{data(ho2mergeMGG)}
\format{
  The format is:
 chr "ho2mergeMGG"
}

\examples{
data(ho2mergeMGG)

}
\keyword{datasets}
