\name{krogan2mergeMGGH}
\alias{krogan2mergeMGGH}
\docType{data}
\title{A data file containing the pre-merged Krogan to the (merged)
  MIPS-GO-Gavin-Ho data}
\description{
   This file contains the pre-processed Krogan data before merging to the
   joint MIPS-GO-Gavin-Ho interactome
}
\usage{data(krogan2mergeMGGH)}
\format{
  The format is:
 chr "krogan2mergeMGGH"
}

\examples{
data(krogan2mergeMGGH)

}
\keyword{datasets}
