\name{localization}
\alias{localization}
\docType{data}
\title{A data file detailing the localization of certain Yeast ORF's}
\description{
  This data file is a named list. The names are the particular cellular
  sub-componenets (locations) of the yeast cell. The entries of the list
  is a character vector of the proteins that localize to that sub-component.
}
\usage{data(localization)}
\format{
  The format is:
 chr "localization"
}
\details{
}
\source{
  
}
\references{
  Huh WK, et al. (2003) Global analysis of protein localization in
  budding yeast. Nature 425(6959):686-91
}
\examples{
data(localization)
}
\keyword{datasets}
