\name{localization2}
\alias{localization2}
\docType{data}
\title{A data set containing each ORF and where is has been localized}
\description{
  The daat-set is a named list each name correspond to each ORF. The
  entries is a character vector of each sub-compnent where the
  corresponding ORF has been experimentally localised
}
\usage{data(localization2)}
\format{
  The format is:
 chr "localization2"
}
\details{

}
\source{

}
\references{
Huh WK, et al. (2003) Global analysis of protein localization in budding
yeast. Nature 425(6959):686-91
}
\examples{
data(localization2)
}
\keyword{datasets}
