\name{mapping2SysG}
\alias{mapping2SysG}
\docType{data}
\title{A example graph of the mapping from IntAct to Systematic Names}
\description{
  A instance of class graph used in the ScISI to show an example of the
  mapping from IntAct to Systematic Names
}
\usage{data(mapping2SysG)}
\format{
  The format is:
 chr "mapping2SysG"
}
\examples{
data(mapping2SysG)

}
\keyword{datasets}
