\name{mappingsG}
\alias{mappingsG}
\docType{data}
\title{A example graph of the mapping from IntAct to Systematic Names}
\description{
   A instance of class graph used in the ScISI to show an example of the
  mapping from IntAct to Systematic Names
}
\usage{data(mappingsG)}
\format{
  The format is:
 chr "mappingsG"
}

}
\examples{
data(mappingsG)

}
\keyword{datasets}
