\name{mips2go}
\alias{mips2go}
\docType{data}
\title{A data file containing the pre-merged GO to the
  MIPS data}
\description{
   This file contains the pre-processed GO data before merging to the
   MIPS data.
}
\usage{data(mips2go)}
\format{
  The format is:
 chr "mips2go"
}

\examples{
data(mips2go)

}
\keyword{datasets}
