\name{nAtMap}
\alias{nAtMap}
\docType{data}
\title{A file containing node attributes}
\description{
  This file contains node attributes used in the rendering of the graph
  ARP 2/3 in the ScISI.Rnw paper.
}
\usage{data(nAtMap)}
\format{
  The format is:
 chr "nAtMap"
}
\examples{
data(nAtMap)

}
\keyword{datasets}
