\name{nonGenes}
\alias{nonGenes}
\docType{data}
\title{Genes found in MIPS which are not gene locus names}
\description{
  A character vector of the genes found in MIPS which are not gene locus names.

  They include: RNA\_TLC1, SNRNA\_NME1, and RNA\_RNASE-P.
}
\usage{data(nonGenes)}
\format{
  The format is:
 chr "nonGenes"
}
\examples{
data(nonGenes)
}
\keyword{datasets}
