\name{nucComp}
\alias{nucComp}
\docType{data}
\title{A data file containing the nuclear complexes}
\description{
  This file is a character vector of the names of those protein
  complexes which have at least 75 percent of its constituent members
  annotated with the nucleus GO term.
}
\usage{data(nucComp)}
\format{
  The format is:
 chr "nucComp"
}
\examples{
data(nucComp)

}
\keyword{datasets}
