\name{subCompM}
\alias{subCompM}
\docType{data}
\title{A matrix of sub-complex summary statistics}
\description{
  This matrix contains the sub-complex summary statistics for the
  ScISI. The rows and columns are indexed by the individual data-set
  from where the protein complex estimates orginated. The (i,j) entry
  details how many protein complexes of the index i are sub-complexes of
  the index j.
}
\usage{data(subCompM)}
\format{
  Matrix
}
\details{
  
}

\examples{
data(subCompM)

}
\keyword{datasets}
