\name{unwanted}
\alias{unwanted}
\docType{data}
\title{GO terms that are parsed but not protein complexes}
\description{
  A character vector of the GO terms parsed but known not to be protein complexes
  in GO.

  The include: GO:0000262, GO:0000228, GO:0000775, GO:0010008, GO:0005792,
  GO:0005768, GO:0005769, GO:0005770, GO:0005777, GO:0005844, and
  GO:0001400.
}
\usage{data(unwanted)}
\format{
  The format is:
 chr "unwanted"
}
\examples{
data(unwanted)
}
\keyword{datasets}
