\name{xtraGO}
\alias{xtraGO}
\docType{data}
\title{A character vector of hand selected GO protein complexes}
\description{
  Hand selected protein complexes that will be used
  to see if any subset should be incorporated into the
  ScISI
}
\usage{data(xtraGO)}
\format{
  The format is:
 chr "xtraGO"
}


\examples{
data(xtraGO)
}
\keyword{datasets}
